Created MfDynmapService with quality & simple implementations
